feat: Applied result pattern to MenuItem endpoints

+++ Application/Features/MenuItem/Commands/DeleteMenuItemByIdCommandHandler.cs
@@ -1,11 +1,13 @@
 ﻿using Application.Interfaces;
 using BLL.Dtos.MenuItem;
 using BLL.Interfaces;
+using DAL.Errors;
+using DAL.SharedKernels;
 using MediatR;
 
 namespace Application.Features.MenuItem.Commands
 {
-    public class DeleteMenuItemByIdCommandHandler : IRequestHandler<DeleteMenuItemByIdCommand, GetMenuItemDto>
+    public class DeleteMenuItemByIdCommandHandler : IRequestHandler<DeleteMenuItemByIdCommand, Result<GetMenuItemDto>>
     {
         private readonly IMenuItemService _menuItemService;
         private readonly IRedisService _redisService;
@@ -16,11 +18,16 @@ namespace Application.Features.MenuItem.Commands
             _redisService = redisService;
         }
 
-        public async Task<GetMenuItemDto> Handle(DeleteMenuItemByIdCommand request, CancellationToken cancellationToken)
+        public async Task<Result<GetMenuItemDto>> Handle(DeleteMenuItemByIdCommand request, CancellationToken cancellationToken)
         {
+            var menuItem = await _menuItemService.DeleteByIdAsync(request.id);
+            if (menuItem == null) 
+                return Result<GetMenuItemDto>.Failure(MenuItemErrors.NotFound);
+
             await _redisService.DeleteDataAsync($"menuItem:{request.id}");
-            return await _menuItemService.DeleteByIdAsync(request.id);
+
+            return Result<GetMenuItemDto>.Success(menuItem);
         }
     }
-    public record DeleteMenuItemByIdCommand(int id) : IRequest<GetMenuItemDto> { }
+    public record DeleteMenuItemByIdCommand(int id) : IRequest<Result<GetMenuItemDto>> { }
 }
